Add files via upload
In this test case:

The test_website_load_time method opens the website URL and measures the load time using JavaScript. Specifically, it calculates the time taken from the start of navigation to the completion of the load event.
You can adjust the max_load_time variable to set the maximum acceptable load time threshold in milliseconds. In this example, it's set to 5000 milliseconds (5 seconds).
The assertLessEqual assertion checks if the actual load time is less than or equal to the specified threshold. If the load time exceeds the threshold, the test will fail with an appropriate error message.